Rename FragmentConst to MainPageConst.
