+++ PodoMarket/app/src/main/java/com/android/example/podomarket/ui/main/MainPagerAdapter.kt
@@ -3,10 +3,10 @@ package com.android.example.podomarket.ui.main
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.android.example.podomarket.ui.main.FragmentConst.ARTICLE
-import com.android.example.podomarket.ui.main.FragmentConst.CHAT
-import com.android.example.podomarket.ui.main.FragmentConst.MYPAGE
-import com.android.example.podomarket.ui.main.FragmentConst.PRODUCT
+import com.android.example.podomarket.ui.main.MainPageConst.ARTICLE
+import com.android.example.podomarket.ui.main.MainPageConst.CHAT
+import com.android.example.podomarket.ui.main.MainPageConst.MYPAGE
+import com.android.example.podomarket.ui.main.MainPageConst.PRODUCT
 
 class MainPagerAdapter(activity: FragmentActivity) : FragmentStateAdapter(activity) {
     override fun getItemCount(): Int = 4
@@ -24,7 +24,7 @@ class MainPagerAdapter(activity: FragmentActivity) : FragmentStateAdapter(activi
     }
 }
 
-object FragmentConst {
+object MainPageConst {
     const val PRODUCT = 0
     const val ARTICLE = 1
     const val CHAT = 2
